Title: 시력 검사 2, Time: 151ms, Memory: 22MB, Status: Passed - Codetree

+++ 260306/시력 검사 2/eye-test-2.java
@@ -4,7 +4,7 @@ public class Main {
 	public static void main(String[] args) {
     	Scanner sc = new Scanner(System.in);
         
-        int a=sc.nextInt();
+        Double a=sc.nextDouble();
 
         if (a>=1.0)System.out.println("High");
         else if (a>=0.5) System.out.println("Middle");
